% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{list_models}
\alias{list_models}
\title{List the available frozen models}
\usage{
list_models()
}
\value{
A data frame with model identifiers and input feature counts.
}
\description{
List the available frozen models
}
\examples{
list_models()
}
